MAINT: reorganize functions in morestats.py related to box-cox transformation
